move user-button
