+++ src/app/layout/layout.module.ts
@@ -8,7 +8,7 @@ import { TransactionsModule } from "../transactions/transactions.module";
 import { LayoutComponent } from "./layout/layout.component";
 import { DrawerComponent } from "./drawer/drawer.component";
 import { ToolbarComponent } from "./toolbar/toolbar.component";
-import { ToolbarUserComponent } from "./toolbar/user-button/user-button.component";
+import { ToolbarUserComponent } from "./user-button/user-button.component";
 import { NotFoundComponent } from "./not-found/not-found.component";
 
 @NgModule({
