+++ src/app/layout/user-button/user-button.component.scss
