+++ src/app/layout/user-button/user-button.component.ts
@@ -1,5 +1,5 @@
 import { Component } from "@angular/core";
-import { AuthService } from "../../../services/auth.service";
+import { AuthService } from "../../services/auth.service";
 
 @Component({
   selector: "app-user-button",
